requirements.txt
# backend/requirements.txt
setuptools>=58.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
paddlepaddle>=2.5.2
paddleocr>=2.6.1.3
pytesseract>=0.3.8
pillow>=8.3.1
pymupdf>=1.18.14
numpy>=1.21.2
python-multipart>=0.0.5
